% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{combine}
\alias{combine}
\title{Combine across effect sizes / multiple outcomes}
\usage{
combine(eff = c(), vi = c(), rho = 0.7)
}
\arguments{
\item{eff}{vector of effect sizes to be combined}

\item{vi}{vector of variances of the effect sizes}

\item{rho}{assumed correlation across effect sizes (applies to all covariances)}
}
\value{
list with combined effect size (mu) and its variance (v)
}
\description{
Combine across effect sizes / multiple outcomes using the strategy proposed by Borenstein et al. (2009, pp.226-229): unweighted average is used for point estimate, which is adequate assuming that very similar variances apply to all effect sizes within a cluster/study. This does a similar job as the aggregate.escalc function in "metafor" using argument weighted=FALSE, but this one may be more practical because its inputs are vectors of effects sizes and variances instead of an "escalc" object.
}
