% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{dppc2}
\alias{dppc2}
\title{Compute Morris (2008) "dppc2"}
\usage{
dppc2(
  cgN,
  cgMpre,
  cgMpost,
  cgSDpre,
  cgSDpost,
  tgN,
  tgMpre,
  tgMpost,
  tgSDpre,
  tgSDpost,
  rho = 0.5
)
}
\arguments{
\item{cgN}{N of control group}

\item{cgMpre}{mean value for control group at pretest}

\item{cgMpost}{mean value for control group at posttest}

\item{cgSDpre}{standard deviation for control group at pretest}

\item{cgSDpost}{standard deviation for control group at posttest}

\item{tgN}{N of treated group}

\item{tgMpre}{mean value for treated group at pretest}

\item{tgMpost}{mean value for treated group at posttest}

\item{tgSDpre}{standard deviation for treated group at pretest}

\item{tgSDpost}{standard deviation for treated group at posttest}

\item{rho}{Assumed correlation for pretest-posttest scores, needed to compute variance}
}
\value{
A list containing multiple outputs.
\describe{
  \item{eff}{The effect size estimate.}
  \item{vi}{The variance of the effect size estimate.}
  \item{vi.0}{The variance of the effect size fixing the effect to zero: needed for examination of publication bias via meta-regression, to make effect size and its variance independent.}
}
}
\description{
Computes "dppc2", that is Cohen's d estimate for pretest-posttest designs, using pooled pretest SD and correction for small samples, as suggested by Morris (2008)
}
