% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{funnelT}
\alias{funnelT}
\title{Enhanced funnel plot}
\usage{
funnelT(fit = NA, petpeese = NA, showStudies = T, xlab = NA, sizetext = 20)
}
\arguments{
\item{fit}{a rma or rma.mv fitted object}

\item{petpeese}{either "petpeese" / "pet-peese" for full method, or "pet" or "peese" only, or empty/anything else for not using meta-regression}

\item{showStudies}{whether it should show the studies with different colors and shapes, default is TRUE (but only works if there are actually random effects)}

\item{xlab}{label to be used on the x axis of the funnel plot, if not specified label is "Effect size"}

\item{sizetext}{font size to be used in plot}
}
\value{
funnel plot drawn with ggplot2
}
\description{
Depicts funnel plot with colors and shapes for studies in multilevel meta-analysis, and possibly plot pet-peese meta-regression if required
}
