% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{hg}
\alias{hg}
\title{Cohen's d with Hedge's correction for small samples}
\usage{
hg(cgN, cgM, cgSD, tgN, tgM, tgSD)
}
\arguments{
\item{cgN}{N of control group}

\item{cgM}{mean value for control group}

\item{cgSD}{standard deviation for control group}

\item{tgN}{N of treated group}

\item{tgM}{mean value for treated group}

\item{tgSD}{standard deviation for treated group}
}
\value{
Cohen's d with Hedge's correction for small samples
}
\description{
Computes Cohen's d with Hedge's correction for small samples
}
