% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{logit2prob}
\alias{logit2prob}
\title{Transform logit values into probabilities}
\usage{
logit2prob(logit = NA)
}
\arguments{
\item{logit}{logit value(s)}
}
\value{
Probability
}
\description{
This function transform logit values into probabilities
}
