% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{mcOmega}
\alias{mcOmega}
\title{Calculate McDonald's Omega for latent factor reliability}
\usage{
mcOmega(fit = NULL, stdLoadings = NULL, absoluteValues = TRUE)
}
\arguments{
\item{fit}{An object fitted with the "cfa" function of the "lavaan" package}

\item{stdLoadings}{Alternatively, a vector of standardized factor loadings (provide a single vector)}

\item{absoluteValues}{Logical value, TRUE if loading signs must be ignored because observed variables can indifferently be positive or negative indicators of the latent factor (default)}
}
\value{
A list of omega coefficients (if a "cfa" object is provided) or a single numerical value representing the omega coefficient (if a vector of standardized factor loadings is provided)
}
\description{
Calculate McDonald's Omega for latent factor reliability from a fitted CFA ("cfa" object) in lavaan, or from a single vector of standardized factor loadings
}
