% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{minimize_transactions}
\alias{minimize_transactions}
\title{Minimize transactions among people to settle bills}
\usage{
minimize_transactions(balances, digits = 2)
}
\arguments{
\item{balances}{A named vector with the amount contributed by each person to the total expense}

\item{digits}{The number of digits to round the amounts of money}
}
\value{
The shortest possible list indicating the amounts owed by whom to whom to settle the bill
}
\description{
Splitwise-like function for minimizing transactions for settling bills
}
