% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{prob2logit}
\alias{prob2logit}
\title{Transform probabilities into logit values}
\usage{
prob2logit(prob = NA)
}
\arguments{
\item{prob}{probability value(s) in [0, 1]}
}
\value{
Logit value
}
\description{
This function transform probabilities into logit values
}
