% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffee_ci.R
\name{toffee_ci}
\alias{toffee_ci}
\title{Make pretty two-sided 95pct confidence intervals}
\usage{
toffee_ci(m, se, cf)
}
\arguments{
\item{m}{A vector of sample means.}

\item{se}{A vector of sample standard errors.}

\item{cf}{A compliment to a link function. For example use
\code{\link[base]{I}} for linear models or \code{\link[base]{exp}} for
a binomial model.}
}
\value{
A vector of strings with formatted confidence intervals.
}
\description{
Make pretty two-sided 95pct confidence intervals
}
\examples{
toffee_ci(c(1, 2, 3), c(.2, .25, .3), I)
# "[0.61, 1.39]" "[1.51, 2.49]" "[2.41, 3.59]"

toffee_ci(c(4, 2, 7), c(.15, .2, .1), exp)
# "[3.71, 4.29]" "[1.61, 2.39]" "[6.8, 7.2]"
}
