% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffee_forest.R
\docType{data}
\name{toffee_forest}
\alias{toffee_forest}
\title{Attributes about 31 trees.}
\format{A data frame with 31 rows and 5 variables:
\describe{
\item{Girth}{Tree diameter in inches.}
\item{Height}{Height in feet.}
\item{Volume}{Volume of timber in cubic feet.}
\item{Branches}{Number of tree branches.}
\item{Healthy}{Health status of tree.}
}}
\usage{
toffee_forest
}
\description{
A dataset containing information about 31 trees, adapted from \link[datasets:trees]{datasets::trees}.
}
\keyword{datasets}
