% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffee_signif.R
\name{toffee_signif}
\alias{toffee_signif}
\title{Specify symbols to represent confidence levels.}
\usage{
toffee_signif(p_values, thresholds = c(0.05, 0.01, 0.001),
  chars = c("*", "**", "***"))
}
\arguments{
\item{p_values}{A numeric vector of p-values.}

\item{thresholds}{Descending p-value thresholds. Values less than or equal to
a threshold will produce the corresponding character.}

\item{chars}{A vector of strings that symbolize a p-value less than the
threshold.}
}
\description{
Specify symbols to represent confidence levels.
}
\examples{

toffee_signif(c(1, .5, .049, .01, .0012, 0))
}
