% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffee_tbl.R
\name{toffee_tbl}
\alias{toffee_tbl}
\title{Create a nicely formatted table from a linear model.}
\usage{
toffee_tbl(model, conf_level = 0.95, digits = 2,
  concat_signif = TRUE, odds_ratio = TRUE, ...)
}
\arguments{
\item{model}{A model object that is compatible with \code{\link[broom:tidy]{broom::tidy()}}.}

\item{conf_level}{The confidence level that will be used for computing the
confidence interval.}

\item{digits}{The number of decimal places that numbers in the table should
be rounded to. For no rounding use \code{Inf}. This argument is provided to
\code{\link[base:round]{base::round()}}. The default value is \code{2}.}

\item{concat_signif}{If \code{TRUE} the resulting data frame will append
significance symbols as generated by \code{\link[=toffee_signif]{toffee_signif()}}. Otherwise the
significance symbols will appear in a separate column called \code{Significant}.}

\item{odds_ratio}{If \code{TRUE} and the family of the model provided is
\code{logit} then the model coefficients and confidence intervals returned in
the resulting table will be odds ratios.}

\item{...}{Arguments that will be passed to \link{toffee_signif}. For example,}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the following columns: the
name of the coefficient in the model (\code{Variable}), the value of the
coefficient (\code{Estimate} or \code{Odds_Ratio}), the
confidence interval for the coefficient (\code{CI}), the p-value for the term,
(\code{p_value}), and optionally symbols representing levels of significance
(\code{Significant}).
}
\description{
Given a general linear model (like from \code{\link[stats:lm]{stats::lm()}} or \code{\link[stats:glm]{stats::glm()}}) this
function makes a table intended for interactive use or publication.
}
\examples{

library(dplyr)
library(toffee)

set.seed(2017-10-15)

tree_data <- trees \%>\%
  mutate(Branches = sample(15:25, size = 31, replace = TRUE)) \%>\%
  mutate(Healthy = Branches > 20) \%>\%
  mutate(Healthy = Healthy + sample(c(0, 1), size = nrow(.), replace = TRUE)) \%>\%
  mutate(Healthy = Healthy > 0)

linear_model <- tree_data \%>\%
  lm(Volume ~ Girth + Height + Branches, data = .)

# Basic usage
linear_model \%>\%
  toffee_tbl()

 # # A tibble: 4 x 4
 #   Variable    Estimate CI               p_value
 #   <chr>          <dbl> <chr>            <chr>
 # 1 (Intercept)   -67.2  [-86.81, -47.59] < 0.01***
 # 2 Girth           4.75 [4.23, 5.27]     < 0.01***
 # 3 Height          0.33 [0.08, 0.59]     0.01*
 # 4 Branches        0.46 [-0.03, 0.95]    0.07

# Using the `chars` argument from toffee_signif
linear_model \%>\%
  toffee_tbl(thresholds = 0.01, chars = "*")

 # # A tibble: 4 x 4
 #   Variable    Estimate CI               p_value
 #   <chr>          <dbl> <chr>            <chr>
 #   1 (Intercept)   -67.2  [-86.81, -47.59] < 0.01*
 #   2 Girth           4.75 [4.23, 5.27]     < 0.01*
 #   3 Height          0.33 [0.08, 0.59]     0.01
 #   4 Branches        0.46 [-0.03, 0.95]    0.07

# Separating the significance symbols into their own column
linear_model \%>\%
  toffee_tbl(concat_signif = FALSE)

logistic_model <- tree_data \%>\%
  glm(Healthy ~ Girth + Height + Branches + Volume, data = .,
      family = binomial())

logistic_model \%>\%
  toffee_tbl()

logistic_model \%>\%
  toffee_tbl(thresholds = 0.01, chars = "*")
}
