% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{try_seed}
\alias{try_seed}
\title{Try different seeds for a particular chunk of code}
\usage{
try_seed(expr, maxrun = 10)
}
\arguments{
\item{expr}{Chunk of code to try, among {} brackets}

\item{maxrun}{The maximum number of times it reruns the code}
}
\value{
Whatever the code chunk produce (include print for text/numbers or plot for figures) plus the seed used
}
\description{
(Re)runs a chunk of code with a different seed every time you press ENTER. Useful if the code includes random number generation. Press ESC to exit
}
