% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{z2SumScore}
\alias{z2SumScore}
\title{Transform z-score(s) into Sum Score(s)}
\usage{
z2SumScore(
  z = NA,
  nItems = 10,
  minResp = 1,
  maxResp = 4,
  itemZmin = -2,
  itemZmax = +2
)
}
\arguments{
\item{z}{A single z-score or a vector of z-scores, plausibly reflecting the "true" latent scores. Must be numeric}

\item{nItems}{The number of items composing the test or questionnaire}

\item{minResp}{The minimum observable response to a single item (for binomial responses 0 is recommended; for a Likert scale in 1-4, should be 1). Must be an integer number}

\item{maxResp}{The maximum observable response to a single item (for binomial responses 1 is recommended; for a Likert scale in 1-4, should be 4). Must be an integer number}

\item{itemZmin}{The item difficulty of the easiest item, on a z-score scale. Other item difficulties will be computed on equal interval between intemZmin and itemZmax}

\item{itemZmax}{The item difficulty of the most difficult item, on a z-score scale}
}
\value{
The "z" vector transformed in a vector of plausible sum scores
}
\description{
Transform z-score(s) into "plausible" Sum Score(s) that could be obtained from a test or questionnaire with a finite number of items with binomial or ordinal responses
}
