% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{minimize_transactions}
\alias{minimize_transactions}
\title{Minimize Transactions for Bill Settlement}
\usage{
minimize_transactions(balances, digits = 2)
}
\arguments{
\item{balances}{A named vector representing each person's contribution to the total expense. People who have contributed no money must have values equal to zero}

\item{digits}{An integer specifying the number of decimal places for rounding transaction amounts. Default is 2}
}
\value{
A list of transactions that minimizes the number of payments needed to settle the shared expense, showing who pays whom and how much
}
\description{
A function that efficiently settle shared expenses by minimizing the number of transactions between contributors, similar to the approach used by Splitwise
}
\examples{
# Define contributions
balances = c(Amber = 120, Julie = 20, Phil = 0, Tommy=0)
# Calculate the minimum transactions to settle
transactions = minimize_transactions(balances)
print(transactions)
}
