% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_utils.R
\name{Polar_json2dataframe}
\alias{Polar_json2dataframe}
\title{Polar json to dataframe}
\usage{
Polar_json2dataframe(file = NA)
}
\arguments{
\item{file}{name of the .json file to process}
}
\value{
a dataframe with all relevant data processed from the .json file
}
\description{
Takes one of 4 type of Polar data output: 1) heart rate (files starting with "247ohr-..."); 2) activity (files starting with "activity-..."); 3) temperature (files starting with "generic-period-..."); 4) PPI (files starting with "ppi_samples_...")
}
