% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffee_tbl.R
\name{toffee_tbl}
\alias{toffee_tbl}
\title{Create a pretty table from a model}
\usage{
toffee_tbl(model)
}
\arguments{
\item{model}{A model object that is compatible with \code{\link[broom]{tidy}}.}
}
\value{
A \code{\link[base]{data.frame}} with the following columns: the
name of the coefficient in the model, the value of the coefficient, the
confidence interval for the coefficient, and the p-value for the term.
}
\description{
Given a general linear model (from \code{lm} or \code{glm}) this function
makes a pretty table intended for communication.
}
\examples{

library(dplyr)

set.seed(2017-10-15)

trees \%>\%
  mutate(Branches = sample(15:25, size = 31, replace = TRUE)) \%>\%
  lm(Volume ~ Girth + Height + Branches, data = .) \%>\%
  toffee_tbl()

# term               Coef               CI p.value
# (Intercept) -67.1978389 [-85.93, -48.46] < 0.01*
# Girth         4.7531349     [4.26, 5.25] < 0.01*
# Height        0.3311483     [0.09, 0.58]   0.01*
# Branches      0.4600241    [-0.01, 0.93]    0.07
}
